+++ activejob-delayed_execution.gemspec
@@ -1,11 +1,11 @@
 lib = File.expand_path('../lib', __FILE__)
 $LOAD_PATH.unshift(lib) unless $LOAD_PATH.include?(lib)
 
-require 'activejob/delayed_execution/version'
+require 'activejob_delayed_execution/version'
 
 Gem::Specification.new do |spec|
   spec.name          = 'activejob-delayed_execution'
-  spec.version       = ActiveJob::DelayedExecution::VERSION
+  spec.version       = ActiveJobDelayedExecution::VERSION
   spec.authors       = ['alpaca-tc']
   spec.email         = ['alpaca-tc@alpaca.tc']
