--- lib/activejob/delayed_execution/delayed_execution_job.rb
@@ -1,9 +1,0 @@
-module ActiveJob
-  module DelayedExecution
-    class DelayedExecutionJob < ActiveJob::Base
-      def perform(object, public_method_name, *arguments)
-        object.public_send(public_method_name, *arguments)
-      end
-    end
-  end
-end
