--- lib/activejob/delayed_execution/version.rb
@@ -1,5 +1,0 @@
-module ActiveJob
-  module DelayedExecution
-    VERSION = '0.1.2'
-  end
-end
