+++ lib/activejob_delayed_execution/version.rb
@@ -1,0 +1,3 @@
+module ActiveJobDelayedExecution
+  VERSION = '0.1.2'
+end
